pyproject.toml
[project]
name = "speaches"
version = "0.1.0"
requires-python = "==3.12.*"
# https://packaging.python.org/en/latest/specifications/version-specifiers/#id5
dependencies = [
    "fastapi>=0.121.1",
    "numpy>=2.3.1",
    "onnxruntime>=1.22.0",
    "pydantic-settings>=2.10.1",
    "pydantic>=2.11.7",
    "python-multipart>=0.0.10",
    "soundfile>=0.13.1",
    "uvicorn>=0.35.0",
    "cachetools>=6.1.0",
    "httpx-ws>=0.7.2",
    "av>=14.5.0",
    "aiortc>=1.13.0",
    "httpx>=0.28.1",
    "openai[realtime]>=2.7.1",
    "gradio>=5.49.1",
    "httpx-sse>=0.4.1",
    "pyyaml>=6.0",
    "opentelemetry-distro>=0.50b0",
    "opentelemetry-exporter-otlp>=1.27.0",
    "opentelemetry-instrumentation-asgi==0.50b0",
    "opentelemetry-instrumentation-asyncio==0.50b0",
    "opentelemetry-instrumentation-logging==0.50b0",
    "opentelemetry-instrumentation-threading==0.50b0",
    "opentelemetry-instrumentation-urllib==0.50b0",
    "opentelemetry-instrumentation-fastapi==0.50b0",
    "opentelemetry-instrumentation-grpc==0.50b0",
    "opentelemetry-instrumentation-httpx==0.50b0",
    "opentelemetry-instrumentation-requests==0.50b0",
    "opentelemetry-instrumentation-urllib3==0.50b0",
    "opentelemetry-instrumentation-openai-v2>=2.1b0",
    "opentelemetry-instrumentation-openai>=0.37.1",
]

[dependency-groups]
dev = [
  {include-group = "docs"},
  {include-group = "lint"},
  {include-group = "test"},
]
docs = [
    "mdx-truly-sane-lists>=1.3",
    "mkdocs-material>=9.6.15",
    "mkdocs-render-swagger-plugin>=0.1.2",
    "mkdocstrings[python]>=0.29.1",
]
lint = [
    "basedpyright>=1.33.0",
    "pre-commit>=4.2.0",
    "ruff>=0.14.4",
]
test = [
    "anyio>=4.9.0",
    "datamodel-code-generator>=0.31.2",
    "pytest-antilru>=2.0.0",
    "pytest-asyncio>=1.0.0",
    "pytest-mock>=3.14.1",
    "pytest-xdist>=3.8.0",
    "pytest>=8.4.1",
    "srt>=3.5.3",
    "webvtt-py>=0.5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function" # this fixes pytest warning
filterwarnings = [
    'ignore::DeprecationWarning'
]
# https://docs.pytest.org/en/stable/how-to/mark.html#registering-marks
markers = [
    "requires_openai",
]

[tool.uv]
required-version = ">=0.8.14"

[tool.uv.workspace]
members = [
    "packages/speaches-cli",
]
